refactor(db): switch from SQLite to Supabase PostgreSQL
The dev SQLite substitute is replaced by the real PostgreSQL backend (spec
target), hosted on Supabase.

- driver: @libsql/client -> postgres (postgres.js); client/migrate use
  drizzle-orm/postgres-js with { ssl: "require", prepare: false } (prepare
  off keeps it compatible with the Supavisor transaction pooler)
- schema: all 7 tables sqlite-core -> pg-core (pgTable/serial/jsonb/boolean/
  timestamptz); json/enum/timestamp columns become native pg types
- activity aggregate: cast sum() to ::int (pg returns sum as a string)
- drizzle.config dialect -> postgresql; DATABASE_URL now required (no SQLite
  default); .env.example + README updated
- migrations regenerated for postgres (old SQLite migration removed)

Verified: build/typecheck/lint pass; migration applied to Supabase; an
end-to-end smoke test (guild config, atomic activity increment + aggregate,
honeypot CRUD, moderation log) passed against the live DB and cleaned up.

+++ packages/db/src/repositories/activity-repository.ts
@@ -89,12 +89,14 @@ export function createActivityRepository(db: Db): ActivityRepository {
     async aggregateByUserBetween(guildId, from, to) {
       const rows = await db
         .select({
+          // Postgres returns sum() as a numeric string; cast to int so these
+          // come back as JS numbers.
           userId: activityDaily.userId,
-          chatCount: sql<number>`coalesce(sum(${activityDaily.chatCount}), 0)`,
-          voiceSeconds: sql<number>`coalesce(sum(${activityDaily.voiceSeconds}), 0)`,
-          imageCount: sql<number>`coalesce(sum(${activityDaily.imageCount}), 0)`,
-          musicCount: sql<number>`coalesce(sum(${activityDaily.musicCount}), 0)`,
-          interactionCount: sql<number>`coalesce(sum(${activityDaily.interactionCount}), 0)`,
+          chatCount: sql<number>`coalesce(sum(${activityDaily.chatCount}), 0)::int`,
+          voiceSeconds: sql<number>`coalesce(sum(${activityDaily.voiceSeconds}), 0)::int`,
+          imageCount: sql<number>`coalesce(sum(${activityDaily.imageCount}), 0)::int`,
+          musicCount: sql<number>`coalesce(sum(${activityDaily.musicCount}), 0)::int`,
+          interactionCount: sql<number>`coalesce(sum(${activityDaily.interactionCount}), 0)::int`,
         })
         .from(activityDaily)
         .where(
